+++ ctor/src/lib.rs
@@ -3,6 +3,7 @@
 #![doc = include_str!("../docs/BUILD.md")]
 //! # ctor
 #![doc = include_str!("../docs/PREAMBLE.md")]
+#![doc = include_str!("../docs/REEXPORT.md")]
 #![doc = include_str!("../docs/GENERATED.md")]
 // Used as part of ctor collection
 #![cfg_attr(