+++ ctor/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this crate will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Documentation
+
+- Documented the `crate_path` override for re-exporting the macro from another
+  crate.
+
 ## [1.0.12] - 2026-07-29
 
 ### Changed